requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
sentence-transformers
google-generativeai
torch
numpy
pydantic